Update Rules.md
Section edited:
1- Rule commands
Added:
You can also duplicate the same trigger on many lines. 
`on power2#state=1 do power1 1 endon`
`on power2#state=1 do RuleTimer1 100 endon`

**Rules have a maximum string length**
If the string of your rule is to long and it reach the limit of the rule length, you will loose the end of it... Always verify the status line returned after the insert in the console to see if the last words of your rule has been well inserted. In the case of your rule length is larger then the limit, split it in an other rule<x>...

2- Rule Variables
     Removed the % who seams not needed, see the example in the : 
          You can test it directly in the console like:
          12:44:12 CMD: var4 = utctime
          12:44:12 RSL: stat/tasmota/RESULT = {"Var4":"1575481472.000"}

3- Major features > Multiple Tasmota commands into a ...
   Multiple Tasmota commands into a IF works if statements separated by `;`. For example:  
  `IF (Mem1==0) Var1 Var1+1; Mem1 1;  Delay 10; Power1 ENDIF`
  
   But not like this:
   `Power1 off; LedPower on; IF (Mem1==0) Var1 Var1+1; Mem1 1 ENDIF; Delay 10; Power1 on`
   
   You should split it in two lines like:
   `on power2#state=1 do Power1 off; LedPower on; endon`
   `on power2#state=1 do IF (Mem1==0) Var1 Var1+1; Mem1 1 ENDIF endon`