Update Sonoff-B1.md (#46)
Removed the line - **NOTE:** The **B1 R2** needs the GPIO0 connected to the GND pad in the middle of the PCB. Otherwise it won't enter flash mode. - as it is not true.

I have flashed at least 8 of these bulbs and not once used the GND in the middle of the PCB. There were a few bulbs that would not enter flash mode unless GPIO 0 was pulled to GND and held for 20 to 25 seconds but usually 5 to 10 worked fine.