Update Sonoff-4CH-Pro.md
I have owned the Sonoff CH4 PRO for 2 years, I immediately found random activations of the buttons. Not knowing the source of the problem I searched online for similar experiences but these attributed the problem either to electrical disturbances and it was recommended to increase buttondebounce or to MQTT retain. I have made several attempts at both hardware and software without success. In the last month after the purchase of an RF door switch I accidentally found the activation of the buttons in conjunction with the sending of the RF code. I searched online and found other similar findings. The problem is that both the buttons and the RF433 commands pass through MCU STM32 so the ESP8285 receives the button activation signal, not knowing if it originated from a physical button or RF command, the Tasmota console shows “APP: Button1 level 1-0”.
After change I tested the device for 1 week and found the complete resolution of random button activation issues.
I think it is useful to inform in advance through this document those who are about to program Sonoff CH4 PRO with tasmota, simplifying the resolution of the problem. Especially since without targeted research online information is misleading.