Separate note about NodeMCU and D1 mini from the list of Serial-to-USB Adapter devices
The formatting makes it look like you can use NodeMCU and D1 mini to flash devices. I bought one and spent hours trying to make it work only to realise the bullet was in fact noting an exception to flashing those devices, instead of using them to do the flashing.