Update EQ3-TRV.md (#872)
The link at the beginning does not work.
It says that tasmota source does not include eq3 compatibility, but it does, as I have compiled from source and it worked.
So I have deleted that line.

But a description of how to compile the source is missing, so I included it. So I have included it.