Update power toggle mapping info
When relays are on the same board as a TM1637, the power toggle will map to whatever the last defined power button is in the WebUI, so turning the relay on and off will also turn your display on and off, and changing Displaydimmer from <12 to >=12 will turn the relay off and on as well.  Thus a final virtual relay needs to be defined to avoid conflict.