New documentation for EQ3-TRV
This new documentation is based on the experience with the current version of the driver:
All commands and fields of the response are descibed as well as the behavior of the TRV in the different modes.
It already includes BLE_EQ3: Fix and enhance the JSON response #14485
Hopefully the markdown format fits to mkdocs material. I have tested this with my own "mkdocs serve", and it is looking good.