simpler english
I had my problems reading the first paragraph and was confused that I2C is enabled in the official releases, but not on custom builds. It was also not enabled by setting USE_MCP230xx

There was a copy paste error in the used devices. I use my MCP23017 with address 0x27