Update example for Wifi watchdog to working state
The current example for the Wifi Watchdog has two problems: one, it requires Ping which isn't included by default; and two, it doesn't actually work.  The `Ping4` command needs to be wrapped in a backlog in order for the ruleset to work properly.  However, we have something more flexible than Ping - WebQuery.  It's also included in Tasmota by default.  This documentation change updates the example to use WebQuery in a way that actually works.