Update DS18x20.md
Add the comment - "Note: Experience has shown that the 4k7 is necessary even for a single DS18x20 sensor".  

If the 4k7 is added then this command is not required so maybe a better change would be to remove the reference to the command and to add a note making the 4k7 mandatory.