Changed 'and' to 'or' for sunset/sunrise compare (#1203)
This wrong example took me some time troubleshooting.... The if statement will never evaluate to `true` by using `and`. An `or` should be used instead.

When i print `;print Time is %time%. Sunrise is at %sunrise%, sunset is at %sunset%.` in the console of my Tasmota device in timezone Europe/Amsterdam and with the correct local time i get:
`Time is 1252.00. Sunrise is at 402.00, sunset is at 1131.00.`