Add description of configuring a switch for autodiscovery.
This method works for me quickly/easily. The default instructions only make something show up in the UI if I use the relay, which is not recommended. And there are much more complicated workarounds described, for example at https://blakadder.com/pir-in-tasmota-integration/ , and a number of threads one can find with people trying to figure out how to do this. So I think it would really help to have this method described in the main PIR docs.