doc: change "IP Port" to "TCP port" in Zigbee.md
There is no such thing as "IP ports". In the TCP/IP stack, IP corresponds to the network level of the OSI model. Consequently, IP uses IP addresses. In the TCP/IP stack, TCP corresponds to the transport stack of the OSI model. To address different services on a single IP address, TCP uses TCP ports.