Correct pin numbers for second serial port (they were swapped). (#1356)
The pin numbers for the second serial hardware port of ESP8266 were swapped in the description.

A description of the difference of Serial Tx/Serial Rx and SerBr Tx/SerBr Rx would be also good.
I tested on a ESP 8266
Module type: Generic (18)
Serial Tx/Rx was working as expected on the second hardware port (console and MQTT), but SerBr did not work for me. What's the purpose of this option?